index: Update the GH pages
Update the page to reflect the project changes.

After https://github.com/Freescale/Documentation/pull/42 the
Documentation is deprecated.

Add the link to the current Yocto Project releases.

Suggest downloading the master branch, so this page is always up to
date.

Remove the Documentation files (pdfs) and the script to update the site.

Signed-off-by: Daiane Angolini <daiane.angolini@foundries.io>